# Use base primitives in `equal_object()`