Related to #643 

- We now use `Rf_length()` rather than `vec_size()` in `equal_object()`
- This effectively treats data frames as lists, which is correct when it is viewed as an "object" (i.e. in the case of `list(<data_frame>)`)
- The previous usage of `equal_scalar()` would now be invalid, because it treats data frames as vectors of rows. So I changed it to a macro approach that treats data frames as lists.

The last point has major performance improvements for lists.

``` r
library(vctrs)
vec_equal_old <- vctrs2::vec_equal

x <- rep("a", 1000000)
lst <- list(x)

bench::mark(
  new = vec_equal(lst, lst),
  old = vec_equal_old(lst, lst)
)
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new          6.06ms   6.48ms     153.     91.2KB        0
#> 2 old         10.67ms  11.39ms      87.5    79.5KB        0
```

This means that data frames with list columns have a ridiculous improvement.

``` r
library(vctrs)
library(nycflights13)

nested <- tidyr::nest(flights, data = c(-year, -month))

nested
#> # A tibble: 12 x 3
#>     year month            data
#>    <int> <int> <list<df[,17]>>
#>  1  2013     1   [27,004 × 17]
#>  2  2013    10   [28,889 × 17]
#>  3  2013    11   [27,268 × 17]
#>  4  2013    12   [28,135 × 17]
#>  5  2013     2   [24,951 × 17]
#>  6  2013     3   [28,834 × 17]
#>  7  2013     4   [28,330 × 17]
#>  8  2013     5   [28,796 × 17]
#>  9  2013     6   [28,243 × 17]
#> 10  2013     7   [29,425 × 17]
#> 11  2013     8   [29,327 × 17]
#> 12  2013     9   [27,574 × 17]

# before
bench::mark(
  vec_equal(nested, nested, na_equal = TRUE)
)
#> # A tibble: 1 x 6
#>   expression                                   min median `itr/sec`
#>   <bch:expr>                                 <bch> <bch:>     <dbl>
#> 1 vec_equal(nested, nested, na_equal = TRUE) 188ms  192ms      5.21
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>

# after
bench::mark(
  vec_equal(nested, nested, na_equal = TRUE)
)
#> # A tibble: 1 x 6
#>   expression                                    min median `itr/sec`
#>   <bch:expr>                                 <bch:> <bch:>     <dbl>
#> 1 vec_equal(nested, nested, na_equal = TRUE) 4.15ms 4.42ms      226.
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```